pyproject.toml
[tool.poetry]
name = "lintify"
version = "0.0.2"
description = ""
authors = ["Titov Anton <webdev@titovanton.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
watchdog = "^4.0.1"
click = "^8.1.7"
pyyaml = "^6.0.2"
pydantic = "^2.8.2"
types-pyyaml = "^6.0.12.20240808"


[tool.poetry.group.dev.dependencies]
mypy = "^1.8"
flake8 = "^7.1.1"
autoflake = "^2.3.1"
flake8-import-order = "^0.18.2"
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
lintify = "lintify.cli:main"

[tool.mypy]
mypy_path = "./lintify"

[tool.flake8]
max-line-length = 80
import-order-style = "pep8"
